+++ utility/classes/FileUploadPaths.php
@@ -1,5 +1,6 @@
 <?php
 
+	namespace Gac\DidYouBuyIt\utility\classes;
 
 	class FileUploadPaths
 	{
